Implemented range checking (skills should be updated accordingly to work properly)

+++ TemalabProject/Assets/Scripts/Model/Characters/Assassin/Skills/BasicAttack_Assassin.cs
@@ -35,7 +35,7 @@ namespace Assets.Scripts.Model.Skills {
             source.TurnStats.ActionPoints--;
         }
 
-        public override double GetRange(Character source) {
+        public override float GetRange(Character source) {
             return 10;
         }
